+++ runtime/src/comm/none/comm-none.c
@@ -49,7 +49,7 @@
 // Chapel interface
 chpl_comm_nb_handle_t chpl_comm_put_nb(void *addr, c_nodeid_t node, void* raddr,
                                        size_t size, int32_t commID,
-                                       int ln, int32_t fn)
+                                       int64_t ln, int32_t fn)
 {
   assert(node == 0);
   memmove(raddr, addr, size);
@@ -58,7 +58,7 @@ chpl_comm_nb_handle_t chpl_comm_put_nb(void *addr, c_nodeid_t node, void* raddr,
 
 chpl_comm_nb_handle_t chpl_comm_get_nb(void* addr, c_nodeid_t node, void* raddr,
                                        size_t size, int32_t commID,
-                                       int ln, int32_t fn)
+                                       int64_t ln, int32_t fn)
 {
   assert(node == 0);
   memmove(addr, raddr, size);
@@ -259,14 +259,14 @@ void chpl_comm_pre_task_exit(int all) { }
 void chpl_comm_exit(int all, int status) { }
 
 void  chpl_comm_put(void* addr, c_nodeid_t node, void* raddr,
-                    size_t size, int32_t commID, int ln, int32_t fn) {
+                    size_t size, int32_t commID, int64_t ln, int32_t fn) {
   assert(node==0);
 
   memmove(raddr, addr, size);
 }
 
 void  chpl_comm_get(void* addr, c_nodeid_t node, void* raddr,
-                    size_t size, int32_t commID, int ln, int32_t fn) {
+                    size_t size, int32_t commID, int64_t ln, int32_t fn) {
   assert(node==0);
 
   memmove(addr, raddr, size);
@@ -275,7 +275,7 @@ void  chpl_comm_get(void* addr, c_nodeid_t node, void* raddr,
 void  chpl_comm_put_strd(void* dstaddr_arg, size_t* dststrides, c_nodeid_t dstnode,
                          void* srcaddr_arg, size_t* srcstrides, size_t* count,
                          int32_t stridelevels, size_t elemSize, int32_t commID,
-                         int ln, int32_t fn)
+                         int64_t ln, int32_t fn)
 {
   assert(dstnode==0);
   put_strd_common(dstaddr_arg, dststrides, dstnode,
@@ -288,7 +288,7 @@ void  chpl_comm_put_strd(void* dstaddr_arg, size_t* dststrides, c_nodeid_t dstno
 void  chpl_comm_get_strd(void* dstaddr_arg, size_t* dststrides, c_nodeid_t srcnode,
                          void* srcaddr_arg, size_t* srcstrides, size_t* count,
                          int32_t stridelevels, size_t elemSize, int32_t commID,
-                         int ln, int32_t fn)
+                         int64_t ln, int32_t fn)
 {
   assert(srcnode==0);
   get_strd_common(dstaddr_arg, dststrides, srcnode,
@@ -301,7 +301,7 @@ void  chpl_comm_get_strd(void* dstaddr_arg, size_t* dststrides, c_nodeid_t srcno
 void chpl_comm_getput_unordered(c_nodeid_t dstnode, void* dstaddr,
                                 c_nodeid_t srcnode, void* srcaddr,
                                 size_t size, int32_t commID,
-                                int ln, int32_t fn)
+                                int64_t ln, int32_t fn)
 {
   assert(srcnode==0);
   assert(dstnode==0);
@@ -309,14 +309,14 @@ void chpl_comm_getput_unordered(c_nodeid_t dstnode, void* dstaddr,
 }
 
 void chpl_comm_get_unordered(void* addr, c_nodeid_t node, void* raddr,
-                             size_t size, int32_t commID, int ln, int32_t fn)
+                             size_t size, int32_t commID, int64_t ln, int32_t fn)
 {
   assert(node == 0);
   memmove(addr, raddr, size);
 }
 
 void chpl_comm_put_unordered(void* addr, c_nodeid_t node, void* raddr,
-                             size_t size, int32_t commID, int ln, int32_t fn)
+                             size_t size, int32_t commID, int64_t ln, int32_t fn)
 {
   assert(node == 0);
   memmove(raddr, addr, size);
